Navigation moved to Header / modify class name for toggle visibility of menu
