+++ src/App.js
@@ -1,11 +1,11 @@
-import Navigation from "./components/Navigation";
+import Header from "./components/Header";
 
 import "./styles/main.scss";
 
 function App() {
   return (
     <div className="App">
-      <Navigation />
+      <Header />
     </div>
   );
 }
