+++ src/components/Header/Navigation/Navigation.js
@@ -1,0 +1,6 @@
+import React from "react";
+import NavList from "./NavList";
+
+export default function Navigation() {
+  return <NavList />;
+}
