+++ src/styles/main.scss
@@ -4,7 +4,7 @@
 @import "navigation.scss";
 
 body {
-  background-color: $persianGreen;
   width: 100%;
-  height: 100vh;
+  height: 100%;
+  background-color: $charcoal;
 }
